@@ -12,23 +12,23 @@
   "private": true,
   "dependencies": {
     "@alyle/ui": "^16.0.1",
-    "@angular/animations": "^17.0.8",
-    "@angular/cdk": "^17.0.5",
-    "@angular/common": "^17.0.8",
-    "@angular/compiler": "^17.0.8",
-    "@angular/core": "^17.0.8",
-    "@angular/fire": "^17.0.1",
-    "@angular/forms": "^17.0.8",
-    "@angular/material": "^17.0.5",
-    "@angular/material-moment-adapter": "^17.0.5",
-    "@angular/platform-browser": "^17.0.8",
-    "@angular/platform-browser-dynamic": "^17.0.8",
-    "@angular/router": "^17.0.8",
-    "@ng-bootstrap/ng-bootstrap": "^16.0.0",
+    "@angular/animations": "^18.1.2",
+    "@angular/cdk": "^18.1.2",
+    "@angular/common": "^18.1.2",
+    "@angular/compiler": "^18.1.2",
+    "@angular/core": "^18.1.2",
+    "@angular/fire": "^18.0.1",
+    "@angular/forms": "^18.1.2",
+    "@angular/material": "^18.1.2",
+    "@angular/material-moment-adapter": "^18.1.2",
+    "@angular/platform-browser": "^18.1.2",
+    "@angular/platform-browser-dynamic": "^18.1.2",
+    "@angular/router": "^18.1.2",
+    "@ng-bootstrap/ng-bootstrap": "^17.0.0",
     "@ngx-translate/core": "^15.0.0",
     "@ngx-translate/http-loader": "^8.0.0",
     "@popperjs/core": "^2.11.6",
-    "@syncfusion/ej2-angular-charts": "^24.2.9",
+    "@syncfusion/ej2-angular-charts": "^26.2.5",
     "body-parser": "^1.20.2",
     "bootstrap": "^5.2.3",
     "chart.js": "^4.3.3",
@@ -50,23 +50,23 @@
     "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^17.0.0",
-    "@angular-devkit/build-angular": "^17.0.9",
-    "@angular/cli": "^17.0.9",
-    "@angular/compiler-cli": "^17.0.8",
-    "@angular/localize": "^17.0.8",
+    "@angular-builders/custom-webpack": "^18.0.0",
+    "@angular-devkit/build-angular": "^18.1.2",
+    "@angular/cli": "^18.1.2",
+    "@angular/compiler-cli": "^18.1.2",
+    "@angular/localize": "^18.1.2",
     "@types/body-parser": "^1.19.2",
     "@types/chart.js": "^2.9.37",
     "@types/hammerjs": "^2.0.42",
     "@types/jasmine": "~5.1.4",
     "@types/node": "^20.11.27",
-    "jasmine-core": "~5.1.2",
+    "jasmine-core": "^5.2.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "^2.1.0",
-    "typescript": "5.4.2",
+    "typescript": "^5.5.4",
     "webpack": "^5.77.0",
     "webpack-cli": "^5.0.1"
   },
