@@ -12,7 +12,7 @@ import { DiscussioHouseRoutingModule } from './discussi/discussio-house/discussi
 import { SelectedFlatService } from './services/selected-flat.service';
 import { DragDropModule } from '@angular/cdk/drag-drop';
 import { CommonModule, DatePipe } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatCardModule } from '@angular/material/card';
@@ -69,90 +69,84 @@ import { StatusMessageService } from './services/status-message.service';
 import { CardsDataService } from './services/user-components/cards-data.service';
 import { PostsComponent } from './pages/posts/posts.component';
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    OurTeamComponent,
-    InformationUserComponent,
-    UserLicenceComponent,
-    ComunPageComponent,
-    LookingComponent,
-    FeedbackComponent,
-    AboutProjectComponent,
-    OpportunitiesComponent,
-    ReportsComponent,
-    NotFoundComponent,
-    NewsComponent,
-    SupportUsComponent,
-    ActTransferComponent,
-    RentalAgreementComponent,
-    ConfirmActionsComponent,
-    HomeComponent,
-    HomePageComponent,
-    TemplateComponent,
-    SharingInfoComponent,
-    NewsLineComponent,
-    StatusMessageComponent,
-    PostsComponent,
-  ],
-  providers: [
-    SelectedFlatService,
-    StatusMessageService,
-    CardsDataService,
-    DatePipe,
-    SharedService,
-    StyleRenderer,
-    LyTheme2,
-    { provide: LY_THEME_NAME, useValue: 'minima-deep-dark' },
-    { provide: LY_THEME, useClass: MinimaLight, multi: true },
-    { provide: LY_THEME, useClass: MinimaDeepDark, multi: true },
-    { provide: LY_THEME, useClass: MinimaDark, multi: true },
-  ],
-  bootstrap: [AppComponent],
-  imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
-    FormsModule,
-    NgbModule,
-    MatIconModule,
-    MatCardModule,
-    DragDropModule,
-    HttpClientModule,
-    MatDatepickerModule,
-    MatNativeDateModule,
-    CommonModule,
-    MatSelectModule,
-    MatDialogModule,
-    MatSliderModule,
-    SharedModule,
-    MatIconModule,
-    RouterModule,
-    MatAutocompleteModule,
-    MatTooltipModule,
-    MatStepperModule,
-    HousingServicesModule,
-    HousingServicesRoutingModule,
-    ReactiveFormsModule,
-    MatRadioModule,
-    MatMenuModule,
-    MatCheckboxModule,
-    HammerModule,
-    MatProgressSpinnerModule,
-    LyImageCropperModule,
-    LySliderModule,
-    LyButtonModule,
-    LyIconModule,
-    LyDialogModule,
-    MatInputModule,
-    MatFormFieldModule,
-    HousingParametersModule,
-    HousingParametersRoutingModule,
-    SearchModule,
-    SearchRoutingModule,
-    DiscussioRoutingModule,
-    DiscussioHouseRoutingModule,
-    AuthModule,
-    AuthRoutingModule,
-  ]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        OurTeamComponent,
+        InformationUserComponent,
+        UserLicenceComponent,
+        ComunPageComponent,
+        LookingComponent,
+        FeedbackComponent,
+        AboutProjectComponent,
+        OpportunitiesComponent,
+        ReportsComponent,
+        NotFoundComponent,
+        NewsComponent,
+        SupportUsComponent,
+        ActTransferComponent,
+        RentalAgreementComponent,
+        ConfirmActionsComponent,
+        HomeComponent,
+        HomePageComponent,
+        TemplateComponent,
+        SharingInfoComponent,
+        NewsLineComponent,
+        StatusMessageComponent,
+        PostsComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        BrowserAnimationsModule,
+        FormsModule,
+        NgbModule,
+        MatIconModule,
+        MatCardModule,
+        DragDropModule,
+        MatDatepickerModule,
+        MatNativeDateModule,
+        CommonModule,
+        MatSelectModule,
+        MatDialogModule,
+        MatSliderModule,
+        SharedModule,
+        MatIconModule,
+        RouterModule,
+        MatAutocompleteModule,
+        MatTooltipModule,
+        MatStepperModule,
+        HousingServicesModule,
+        HousingServicesRoutingModule,
+        ReactiveFormsModule,
+        MatRadioModule,
+        MatMenuModule,
+        MatCheckboxModule,
+        HammerModule,
+        MatProgressSpinnerModule,
+        LyImageCropperModule,
+        LySliderModule,
+        LyButtonModule,
+        LyIconModule,
+        LyDialogModule,
+        MatInputModule,
+        MatFormFieldModule,
+        HousingParametersModule,
+        HousingParametersRoutingModule,
+        SearchModule,
+        SearchRoutingModule,
+        DiscussioRoutingModule,
+        DiscussioHouseRoutingModule,
+        AuthModule,
+        AuthRoutingModule], providers: [
+        SelectedFlatService,
+        StatusMessageService,
+        CardsDataService,
+        DatePipe,
+        SharedService,
+        StyleRenderer,
+        LyTheme2,
+        { provide: LY_THEME_NAME, useValue: 'minima-deep-dark' },
+        { provide: LY_THEME, useClass: MinimaLight, multi: true },
+        { provide: LY_THEME, useClass: MinimaDeepDark, multi: true },
+        { provide: LY_THEME, useClass: MinimaDark, multi: true },
+        provideHttpClient(withInterceptorsFromDi()),
+    ] })
 export class AppModule { }
