@@ -5,7 +5,7 @@ import { LoginComponent } from './login/login.component';
 import { RegistrationComponent } from './registration/registration.component';
 import { ChangePasswordComponent } from './change-password/change-password.component';
 import { DragDropModule } from '@angular/cdk/drag-drop';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { MatDialogModule } from '@angular/material/dialog';
@@ -30,46 +30,39 @@ import { MatRadioModule } from '@angular/material/radio';
 import { MatSliderModule } from '@angular/material/slider';
 import { MatStepperModule } from '@angular/material/stepper';
 
-@NgModule({
-  declarations: [
-    AuthComponent,
-    LoginComponent,
-    RegistrationComponent,
-    ChangePasswordComponent,
-  ],
-  providers: [
-    DatePipe
-  ],
-  imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
-    NgbModule,
-    DragDropModule,
-    HttpClientModule,
-    MatIconModule,
-    MatDialogModule,
-    FormsModule,
-    ReactiveFormsModule,
-    CommonModule,
-    SharedModule,
-    RouterModule,
-    MatAutocompleteModule,
-    MatSelectModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatTooltipModule,
-
-    MatCardModule,
-    MatDatepickerModule,
-    MatNativeDateModule,
-    MatSliderModule,
-    MatStepperModule,
-    MatRadioModule,
-    MatMenuModule,
-    MatCheckboxModule,
-    HammerModule,
-    MatProgressSpinnerModule,
-  ]
-})
+@NgModule({ declarations: [
+        AuthComponent,
+        LoginComponent,
+        RegistrationComponent,
+        ChangePasswordComponent,
+    ], imports: [BrowserModule,
+        BrowserAnimationsModule,
+        NgbModule,
+        DragDropModule,
+        MatIconModule,
+        MatDialogModule,
+        FormsModule,
+        ReactiveFormsModule,
+        CommonModule,
+        SharedModule,
+        RouterModule,
+        MatAutocompleteModule,
+        MatSelectModule,
+        MatFormFieldModule,
+        MatInputModule,
+        MatTooltipModule,
+        MatCardModule,
+        MatDatepickerModule,
+        MatNativeDateModule,
+        MatSliderModule,
+        MatStepperModule,
+        MatRadioModule,
+        MatMenuModule,
+        MatCheckboxModule,
+        HammerModule,
+        MatProgressSpinnerModule], providers: [
+        DatePipe,
+        provideHttpClient(withInterceptorsFromDi())
+    ] })
 
 export class AuthModule { }
